редактирует заглушку для фильтров
